style: remove functions from long and hard to read file
